adding tests and refactoring

+++ src/main/java/com/gympass/kartrace/domain/Lap.java
@@ -1,4 +1,4 @@
 package com.gympass.kartrace.domain;
 
-public record Lap (int number, double timeSeconds, double averageVelocity){
+public record Lap (int number, Float timeSeconds, Float averageVelocity){
 }
